Rename game module to chess
